requirements.txt
openpyxl
pandas
flask
sentence_transformers
faiss-cpu
langdetect
transformers
fasttext-wheel
deepl
requests
